# frontend/openapi/return-platform.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Return Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/data-console/v1/browser/{engine}/{asset_id}/records": {
      "get": {
        "tags": [
          "Data Browser"
        ],
        "summary": "Get Records",
        "operationId": "get_records_data_console_v1_browser__engine___asset_id__records_get",
        "parameters": [
          {
            "name": "engine",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engine"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_dict__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/data-console/v1/browser/{engine}/{asset_id}/records/{record_id}": {
      "get": {
        "tags": [
          "Data Browser"
        ],
        "summary": "Get Record",
        "operationId": "get_record_data_console_v1_browser__engine___asset_id__records__record_id__get",
        "parameters": [
          {
            "name": "engine",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engine"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          },
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_dict_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta"
        ],
        "title": "APIResponse[dict]"
      },
...
      "APIResponse_list_dict__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta"
        ],
        "title": "APIResponse[list[dict]]"
      },
...
      "FreshnessStatus": {
        "type": "string",
        "enum": [
          "LIVE",
          "CACHED",
          "STALE"
        ],
        "title": "FreshnessStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          },
          "page_size": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Page Size",
            "default": 50
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PageMeta"
      },
      "ResponseMeta": {
        "properties": {
          "schema_version": {
            "type": "string",
            "pattern": "^\\d+\\.\\d+$",
            "title": "Schema Version",
            "default": "1.0"
          },
          "request_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessStatus",
            "default": "LIVE"
          },
          "partial": {
            "type": "boolean",
            "title": "Partial",
            "default": false
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningMeta"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WarningMeta": {
        "properties": {
          "source": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Source"
          },
          "code": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "code",
          "message"
        ],
        "title": "WarningMeta"
      },
...
    }
  }
}